+++ scaffan/image.py
@@ -163,6 +163,21 @@ class AnnotatedImage:
         self.titles = scan.annotation_titles(self.annotations)
         return self.annotations
 
+    def get_view(self, center=None, level=0, size=None, location=None):
+        view = View(anim=self, center=center, level=level, size=size, location=location)
+        return view
+
+    def get_view_on_annotation(self, annotation_id=None, level=2, boundary_px=10, show=False):
+        annotation_id = self.get_annotation_id(annotation_id)
+        center, size = self.get_annotations_bounds_px(annotation_id)
+        region_size = ((size / self.openslide.level_downsamples[level]) + 2 * boundary_px).astype(int)
+        view = self.get_view(center=center, level=level, size=region_size)
+        if show:
+            view.region_imshow_annotation(annotation_id)
+
+        return view
+
+
     def set_region(self, center=None, level=0, size=None, location=None):
 
         if size is None:
@@ -183,7 +198,7 @@ class AnnotatedImage:
         scan.adjust_annotation_to_image_view(self.openslide, self.annotations,
                                              center, level, size)
 
-    def __select_region(self, i):
+    def get_annotation_id(self, i):
         if type(i) is str:
             i = self.titles[i][0]
         return i
@@ -196,7 +211,7 @@ class AnnotatedImage:
         :param boundary_px:
         :return:
         """
-        i = self.__select_region(i)
+        i = self.get_annotation_id(i)
         center, size = self.get_annotations_bounds_px(i)
         region_size = ((size / self.openslide.level_downsamples[level]) + 2 * boundary_px).astype(int)
         self.set_region(center=center, level=level, size=region_size)
@@ -204,7 +219,7 @@ class AnnotatedImage:
             self.region_imshow_annotation(i)
 
     def get_annotations_bounds_px(self, i=None):
-        i = self.__select_region(i)
+        i = self.get_annotation_id(i)
         if i is not None:
             anns = [self.annotations[i]]
 
@@ -234,12 +249,12 @@ class AnnotatedImage:
         if i is None:
             anns = self.annotations
         else:
-            i = self.__select_region(i)
+            i = self.get_annotation_id(i)
             anns = [self.annotations[i]]
         scan.plot_annotations(anns, in_region=True)
 
     def get_annotation_region_raster(self, i):
-        i = self.__select_region(i)
+        i = self.get_annotation_id(i)
         polygon_x = self.annotations[i]["region_x_px"]
         polygon_y = self.annotations[i]["region_y_px"]
         polygon = list(zip(polygon_y, polygon_x))
@@ -295,9 +310,14 @@ class View:
         self.region_level = level
         self.region_pixelsize, self.region_pixelunit = self.get_pixel_size(level)
         import copy
-        copy.deepcopy(self.anim.annotations)
+        self.annotations = copy.deepcopy(self.anim.annotations)
         scan.adjust_annotation_to_image_view(self.anim.openslide, self.annotations,
                                              center, level, size)
+
+    def get_pixel_size(self, level=None):
+        if level is None:
+            level = self.region_level
+        return self.anim.get_pixel_size(level)
 
     def get_region_image(self, as_gray=False):
         imcr = self.openslide.read_region(
@@ -312,21 +332,29 @@ class View:
         plt.imshow(region)
         self.plot_annotations(i)
 
-    def coords_region_px_to_global_px(self, points_region_px):
+    def coords_glob_px_to_view_px(self, points_glob_px):
+        px_factor = self.anim.openslide.level_downsamples[self.region_level]
+
+        x_px = (points_glob_px[0] - self.region_location[0]) / px_factor
+        y_px = (points_glob_px[1] - self.region_location[1]) / px_factor
+
+        return x_px, y_px
+
+    def coords_view_px_to_glob_px(self, points_view_px):
         """
-        :param points_region_px: [[x0, x1, ...], [y0, y1, ...]]
+        :param points_view_px: [[x0, x1, ...], [y0, y1, ...]]
         :return:
         """
 
-        px_factor = self.openslide.level_downsamples[self.region_level]
-        print(px_factor)
-        x_px = self.region_location[0] + points_region_px[0] * px_factor
-        y_px = self.region_location[1] + points_region_px[1] * px_factor
+        px_factor = self.anim.openslide.level_downsamples[self.region_level]
+        # print(px_factor)
+        x_px = self.region_location[0] + points_view_px[0] * px_factor
+        y_px = self.region_location[1] + points_view_px[1] * px_factor
 
         return x_px, y_px
 
     def get_annotation_region_raster(self, i):
-        i = self.__select_region(i)
+        i = self.anim.get_annotation_id(i)
         polygon_x = self.annotations[i]["region_x_px"]
         polygon_y = self.annotations[i]["region_y_px"]
         polygon = list(zip(polygon_y, polygon_x))
@@ -343,3 +371,22 @@ class View:
         region = self.get_region_image()
         plt.imshow(region)
         self.plot_annotations(i)
+
+    def plot_annotations(self, i=None):
+        if i is None:
+            anns = self.annotations
+        else:
+            i = self.anim.get_annotation_id(i)
+            anns = [self.annotations[i]]
+        scan.plot_annotations(anns, in_region=True)
+
+    def get_region_location_by_center(self, center, level, size):
+        return get_region_location_by_center(self.anim.openslide, center, level, size)
+
+    def get_region_image(self, as_gray=False):
+        imcr = self.anim.openslide.read_region(
+            self.region_location, level=self.region_level, size=self.region_size)
+        im = np.asarray(imcr)
+        if as_gray:
+            im = skimage.color.rgb2gray(im)
+        return im