+++ scaffan/changelog.md
@@ -7,3 +7,4 @@
 * [new] Export annotation to json
 * [new] Convert annotation to json automatically if necessery
 * [new] Get mask from annotation
+* [new] Views can be used for looking on one annotated image